requirements.txt
streamlit
google-generativeai
pandas
plotly
requests
pytz
psycopg2-binary